pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chakload-cli"
version = "1.0.0"
description = "Advanced CLI tool for load testing web applications and Telegram bots with multiple frameworks"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "ChakLoad Team", email = "chakload@example.com"}]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: System :: Benchmark",
    "Topic :: Utilities"
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "PyYAML>=6.0",
    "inquirer>=3.1.0",
    "loguru>=0.7.0",
    "requests>=2.28.0"
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0"
]

[project.scripts]
chakload = "loadtester.__main__:main"

[project.urls]
Homepage = "https://github.com/loadtester/loadtester-cli"
Repository = "https://github.com/loadtester/loadtester-cli"
Documentation = "https://loadtester.readthedocs.io/"

[tool.setuptools.packages.find]
where = ["."]
include = ["loadtester*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]